--- sml_small/thousand_pounds/.coveragerc
@@ -1,2 +1,0 @@
-[run]
-omit = tests/*
